+++ src/main/java/sh/ball/audio/effect/AnimationType.java
@@ -1,16 +1,20 @@
 package sh.ball.audio.effect;
 
 public enum AnimationType {
-  STATIC("Static"),
-  SEESAW("Seesaw"),
-  LINEAR("Linear"),
-  FORWARD("Forward"),
-  REVERSE("Reverse");
+  STATIC("Static", "Static"),
+  SINE("Sine", "Sine"),
+  SQUARE("Square", "Square"),
+  SEESAW("Seesaw", "Seesaw"),
+  LINEAR("Linear", "Triangle"),
+  FORWARD("Forward", "Sawtooth"),
+  REVERSE("Reverse", "Reverse Sawtooth");
 
   private final String name;
+  private final String displayName;
 
-  AnimationType(String name) {
+  AnimationType(String name, String displayName) {
     this.name = name;
+    this.displayName = displayName;
   }
 
   public static AnimationType fromString(String name) {
@@ -22,8 +26,12 @@ public enum AnimationType {
     return null;
   }
 
+  public String getName() {
+    return name;
+  }
+
   @Override
   public String toString() {
-    return name;
+    return displayName;
   }
 }